Flesh out park command test
