+++ composer.json
@@ -18,7 +18,8 @@
             "cli/includes/compatibility.php",
             "cli/includes/facades.php",
             "cli/includes/helpers.php",
-            "tests/UsesNullWriter.php"
+            "tests/UsesNullWriter.php",
+            "tests/BaseApplicationTestCase.php"
         ],
         "psr-4": {
             "Valet\\": "cli/Valet/"
