+++ tests/CliTest.php
@@ -5,37 +5,11 @@ use Symfony\Component\Console\Tester\ApplicationTester;
 /**
  * @requires PHP >= 8.0
  */
-class CliTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
+class CliTest extends BaseApplicationTestCase
 {
-    use UsesNullWriter;
-
-    public function setUp(): void
-    {
-        $this->prepTestConfig();
-        $this->setNullWriter();
-    }
-
-    public function prepTestConfig()
-    {
-        require_once __DIR__.'/../cli/includes/helpers.php';
-
-        if (Filesystem::isDir(VALET_HOME_PATH)) {
-            Filesystem::rmDirAndContents(VALET_HOME_PATH);
-        }
-
-        Configuration::createConfigurationDirectory();
-        Configuration::writeBaseConfiguration();
-    }
-
     public function test_park_command()
     {
-        if (! getenv('CI')) {
-            // $this->markTestSkipped('This test is only run on CI.');
-        }
-
-        $application = require __DIR__.'/../cli/app.php';
-        $application->setAutoExit(false);
-        $tester = new ApplicationTester($application);
+        [$app, $tester] = $this->appAndTester();
 
         $tester->run(['command' => 'park', 'path' => './tests/output']);
 
@@ -46,6 +20,9 @@ class CliTest extends Yoast\PHPUnitPolyfills\TestCases\TestCase
             $tester->getDisplay()
         );
 
-        // @todo Test actual output, I presume.
+        $paths = data_get(Configuration::read(), 'paths');
+
+        $this->assertEquals(1, count($paths));
+        $this->assertEquals('./tests/output', reset($paths));
     }
 }
